Check for pre-initialized gv_name.
Need to think a bit more about what the expectations are for bview_init.
libtclcad appears to be assuming it doesn't do any memory allocations -
gv_name vls init is handled at commands.c:4577, before bview_init is
called.  callbacks ptbl is also handled separately and before bview_init
is called.

In principle it would seem to be cleaner to have a bview_init and a
corresponding bview_free that takes care of all of the internal
container setup and tear-down...