Rough in an irect subcommand for view faceplate
The interactive rectangle and angle-distance cursor are both faceplate
elements in that they are defined in the HUD (Head-Up Display), but they
have their own GED commands (rect and adc) to manage them.

Eventually for consistency we'll want both adc and irect subcommands
defined, so the command structure offers consistent control over the
view elements in an organized way.  The "view faceplate irect" command
will be more focused than GED's rect - it will be up to other logic, for
example, to utilize the rectangle dimensions for operations like
raytracing or zooming.