Separate the TIG plotting routines from plot3.h
The TIG (Terminal-Independent Graphics) plotting routines in plot3.h and
libbn appear to be used in only one place in the code - rtscale.  Break
them into a subdirectory and a separate header to make it easier to
understand which logic constitutes the "main" plot3 logic used by the
majority of BRL-CAD's code.