Add fb mode to hash.
Noticed it to an extent with the standalone view, but the debug messages
make it obvious in qdm - something about the fb drawing is causing
continual Qt events to be triggered.  I thought maybe it was something
about how I had set up the stand-alone window, but it also happens in
qdm which does not have the issue until fb drawing is enabled.  Will
need to figure out what the deal is there.

Would still prefer to get a texture based visualization going for the
fb... tie visual in isst proves it is possible, just need to figure out
the details.