Roughing in ability to get text back from rt
Starting with https://github.com/juangburgos/QConsoleListener and
modding it to suit our particular use case.  Checkpointing a state where
we can get the rt text output to print in the Qt console, although it's
not detaching properly after the window is closed and the prompt is not
properly restored yet.