See if adding m_notifier as a context avoids msgs
Per a note in this article, try supplying the QSocketNotifier as a
context object to the connection:
https://medium.com/genymobile/how-c-lambda-expressions-can-improve-your-qt-code-8cd524f4ed9f

We were getting warnings like the following:

QSocketNotifier: Invalid socket 40 and type 'Read', disabling...

These happened after the raytracing window was closed, and the notifier
no longer had a source.  Not 100% sure yet if supplying m_notifier is
the correct answer, but so far even removing the warning suppression the
above messages do not seem to occur.