Have the last process listener do cleanup
Restructure the cleanup callbacks slightly.  We key off of the value -1
(which is not a valid I/O channel type) to do the final cleanup, and
have that call made by the console's listener cleanup routine.  That
routine checks for listeners related to the process, and doesn't do the
final call unless all of the other listeners related to that process are
also removed.