Queue the before-prompt printing from a subprocess
The insertion methods that work are also slow, so rather than doing it
every time we get a notification of output we throttle based on timing.
Every 10th of a second, write all accumulated output - otherwise, add
it to an internal QString for later printing.

Because we finalize the subprocess logic already post-listener shutdown,
use the same mechanism to also make sure any queued-but-unprinted text
is written out at the end of the subprocess cycle.