Refactor libtclcad fbserv down into libdm
Next incremental step is to get existing fbserv mechanism working for
rt rendering in qtgl dm's embedded fb.  To set the stage for figuring
out how to do that, take the libtclcad version of the fbserv logic and
abstract the Tcl specific portions into callback functions.  This does
three things:  1)  allows us to avoid duplicating shared logic 2) makes
it clear which portions of this do currently need Tcl, and 3) makes the
logical flow of the event-based listening mechanisms clearer.

We are going to have to in some fashion replace the Tcl listeners with
Qt equivalents - this should make it clearer what specifically we need
to accomplish with those equivalents.