Start trying to wire up Tcl based fbserv in Qt
Before we try the Qt version of this, let's see if we can get the
existing Tcl version working more or less the way it works in Archer.
This gets as far as launching things, but communication doesn't work.
Note that this won't work on Windows regardless, since we're not
providing the interp to the data structures.