Experiment with pkg comm using stdin/stdout
Ideally I'd like to be able to send MGED style framebuffer updating over
a stdout pipe rather than TCP sockets - stash relevant experimental code
testing this out with a variation on the libpkg example.  On Linux at
least the indications are this is possible - we'll have to see what
Windows does.