Remove stdio test - not portable
In this form, the libpkg stdio test isn't going to work on Windows.  The
I/O mechanisms on that platform are too fundamentally different.

May be worth thinking about whether defining a way to provide
alternative functions at various points in the pkg logic flow might
generalize it to different forms of IPC...