Wait for subprocesses in non-interactive mode
Report from user that a script doing a raytrace wasn't producing a file.
Sean suspected that MGED was quitting too early and killing the
processes run by the script, it looks like that was correct.

If we're in script mode, wait for subprocesses to finish up before we
quit MGED.