Checkpoint work towards removing RTG.rtg_vlfree
Studying the vlfree logic, it looks like we may be able to replace most
of the RTG.rtg_vlfree global instances with a list associated with the
database instance.  db_open and db_close can then take care of work that
otherwise the calling code had to know to do (the infamous
BU_VLS_INIT(&RTG.rtg_vlfree) needed for working drawing/plotting
routines.)  This commit doesn't get rid of all rtg_vlfree uses in the
code, but it should be a significant step in that direction.

Need to look to see if we have any situations where drawing routines are
looking to use RTG.rtg_vlfree without an open dbip.  If so we'll need have
those cases use the view's vlfree list instead.