Begin shifting qged towards using QtCADView
Getting close the point where it makes sense to begin the migrate from
the stand-alone qdm test harness to qged, which has the raw material for
some of the additional widgets we will be wanting for application
development.  May want to restructure the qged app and widgets -
something to explore.