More experimenting with QTreeView settings
Unless we're going to display other columns we don't need the header,
but if we hide it the line wrapping behavior of the item drawing is
strange.  Set its height to 1 pixel, so we still get the correct
horizontal scrolling behavior.

Add a layout changed emit to the resizing - this seems to be enough to
trigger the horizontal scrollbar update even when we haven't selected a
tree item.