Need another design run at QAccordionWidget
What's there isn't behaving when we do things like make the window
large, and some of the internal widget resizing is bizarre. Need to take
a step back on this widget and reconsider the behavior we need.  Might
be better to do something like make a QWidget containing a QGridLayout
and roll our own behavior rather than trying to shoehorn QSplitter into
doing this.  QSplitter's widget resizing behavior isn't exactly what
we need for this case (close, but we want to relocate widgets down the
stack rather than resize the one immediately below the spacer being
moved...)