Refactor brlman to ready it for a Qt mode
Rather than implement another copy of the man page viewer logic, see if
we can instead mod brlman to support a Qt display in addition to its
command line and Tk modes.  Eventually the Qt mode should replace the
Tk version, and if Qt is enabled the Tk mode will be disabled.

This commit just sets the stage for Qt GUI work, and doesn't yet include
the interface.