Checkpoint a functional plugin load
This is a bit cheap and probably not the right long term solution, but
it does function so checkpointing.  Basically we turn the qged logic
into a library, which can be linked by plugins so they know what they
need to know about the application structures in order to hook in to
palettes, signals/slots, etc.

The "right" way is probably to recast all of this in terms of Qt's
plugin infrastructure, but it's not clear yet how much work that is:
https://doc.qt.io/qt-5/plugins-howto.html#the-low-level-api-extending-qt-applications