Quiet checkbox signal emits when syncing to view
Whoops - when programmatically setting checkboxes, don't emit the view
updating signals - this is not user action in the GUI, and we don't want
it registering as such.

Manifested in a weird error setting framebuffer to overlay mode - the
first checkbox setting (fb on) triggered a view update based on the
checkbox signal, which then updated the view using the existing settings
(no overlay flag set) and stomped the value supplied to the original
view command. Doh.

The fix is to block signals while we're syncing the GUI elements to
reflect the bview state.  This is a good lesson for future work of this
type - if manipulating GUI elements to reflect application state, be
careful to disable signals that are intended to update the view state.
In the GUI updating scenario we already have the view state, and don't
want it changed.

This also illustrates the utility of being able to track through these
issues in gdb - breakpoints were a big help in figuring out what was
happening here.