Start thinking about view/tool interactions.
So far the way the mouse interacts with the view has not changed as a
consequence of tool selection.  That's going to need to change for the
more advanced tools.  View polygon interactions are a case in point, and
will serve as the test case.

Initial thought is that the QToolPaletteElement will supply an event
filter, whose job will be to in turn call a function (if supplied by the
tool internals to QToolPaletteElement) that does the necessary event
handling specific to the tool behavior.  That event filter will then be
installed on (and removed from) the view (or views in quad mode) every
time the tool selections or active tool palettes change.  We may want to
also support the "Esc" key always clearing the view filter.

(The CADApp::tree_update slot may want to become CADAPP::palette_update
instead to handle both tree display and view event filter updates.  When
switching between palette buttons we won't be triggering that callback
and will need to adjust the view filter in response to the palette
element toggling (needs thought), but when we're switching from one
palette to another (i.e. view to instance edit) then we won't
necessarily have an internal palette event to make the change.)