Stash some early thoughts on fully custom views
Sketch editor (and any similar GUI elements) will need to supplant the
standard view temporarily with custom drawing.  Rough in a way to do
this, although it will be a while before we get to actually porting the
sketch editor to figure out how to make this work for real.  (It's not
needed for polygons, which use custom interactions in a standard BRL-CAD
scene view.)