Start looking at exposing polylines as view objs
With the newer Clipper, in principle we can also operate on non-closed
polygon shapes.  Not entirely sure yet if that's desirable, but go ahead
and look at the practicality of exposing it via the GUI.  Even if we
go ahead and disable it, at least the underlying support will be there
if we do want/need it.