Checkpoint work on splitting mod into its own tool
I think there is a better way to go about this, or at least a less
dangerous way.  What we need is for the polygon list to be updated every
time we select the mod tool.  Right now we're triggering on a signal
from the app's gui view update slot, which is in turn passed back down
to the elements.  This approach functions, but has a lot of potential
for problems - if any of the element slots that get called by that
signal in turn change the view, that will generate another gui view
change, and so on - hello infinite loop.

Another possibility might be to introduce a signal at the QToolPalette
level that is provided to the element when it is unhidden.  That would
be local, without the dangerous interconnections this approach entails.
That's the next experiment, but stashing this anyway so we can refer to
it if needed - however dangerous it may be, it does function.