Start assembling a widget to hold polygon settings
Both creation and modification will need to set some of the same values,
so it makes sense to define a common widget that can be used by both.
Creation doesn't currently use signals, since all applications of the
settings are associated with events, but when we get to modification
that will change.