Fixes for import polygon handling and naming.
The polygon update routine does too much when reading a sketch if we
assign the circle/square/etc. types, so just call the vlist update
instead.  Also, pointers to QString data aren't stable - make a
copy immediately if we need to refer to it later.

The view we calculate may not correspond to the original polygon
creation view, so preferentially stash and restore the original view
using the attributes.  This allows for a viewsnap operation on a
polygon imported from a sketch, if the sketch was originally created by
the polygon export routine.