Stash early thoughts on db tree status check
We need a way for the tree view to know when it needs to update (and how
to update without losing open state, but that's another question).  In
principle we should be able to detect this from the db_i itself, but
doing it without prohibitive cost may be tricky.

This code is a start, but will crash because the updated dbip and the
stored internals in the tree model will be out of sync.  Ideally we
wouldn't have to store the internals, but as the tree state lives in the
comb trees we need some reasonably performant way to get at it.

Will need to revisit the tree model/view with fresh eyes to think about
how best to represent what we will need here.