Start working on a way to calculate vZ values
We need some way to set data_vZ to something physically meaningful in
the scene.  One logical candidate is to base it on closest and furthest
points in existing scene scene objects.

This function takes a scene object, iterates over its vlist, and finds
either the closest or furthest point to the view plane.  That point is
then used as the basis for a data_vZ value.

Should also add a way to do this for an arbitrarily specified x,y,z
point in space.