Stub in a skeleton for a Qt .g model for editing
When the original QtCADTree model was put together, we didn't add the
complexity of needing to edit the model in response to edits, either
from the view or from the .g file (beyond completely rebuilding it from
scratch.)  Even without that factor, the resulting code was fairly
complicated.  (Much of that complexity is inherent to the problem, but
even so we may be able to do better in a second pass...)

This stubs in an empty framework for another attempt at building a
QAbstractItemModel for a .g file, this time with an eye towards also
supporting editing down the road.