Add callback for use with db_update_nref
The struct directory pointer contains much of the information we need
for interfaces, but it has no knowledge of parent child relationships.
Rather than the initial QtCADTree solution of (re)creating the logic to
discover those relationships, set up to piggyback off of the work we're
already having to do anyway with db_update_nref.

db_update_nref must be called to ensure d_nref counts in the directory
structs are correct after edits are made to the hierarchy, but (at the
moment) ALL it does is update those reference counts - it does not
retain any of the parent/child knowledge it must unpack in order to
update those counts.

However, that information IS present in the db_update_nref logic when
the counts are updated.  We take advantage of that by adding a callback
to be invoked at the points where incrementing takes place - it is a
those points a parent/child relationship has been identified.  The
application can thus make whatever arrangements it needs to make, and
the only cracking of the internal comb structure done (as well as
db_lookup for the children) is the one that had to be done anyway for
db_udpate_nref.

This is just the parent/child directory pointers - we may also want to
pass back a pointer to the tr_l.tl_name string, since a hierarchy
display may want to show the name even if the dp does not exist...