Add more notes about QgModel setup.
The generality of the comb structures is going to make this a challenge,
since there's nothing in a data sense preventing combs from including
multiple duplicate members.  Indeed, if that happens we would really
want the GUI to be able to display it intelligently, since that would
offer one way for the user to correct the situation.  Unfortunately,
that is likely to seriously complicate the behind the scenes logic for
the model...