Note the last pre-cherrypicking commit to RELEASE
Not really sure where to put this yet, or even if it is truly necessary,
but we will need to do a cherry-pick patch release to pull some main
changes into RELEASE without a full merge, and I'm seeing warnings that
this will complicate eventual full merges of main into release.

To make sure it's clear where the cherry picking process started, make
an explicit record of the last RELEASE commit made prior to the start of
the process.