More thinking about QgModel
I'm beginning to wonder if we might end up doing most of the QgInstance
work in the per-object callbacks, rather than the nref updating.  We'll
still need the latter for tops updating, most likely (since it's only
after that function is run that we know the correct tops answers) and it
may still be the most convenient way to do the initial QgInstance
population when opening the database for the first time, but once
populated the QgInstances are local to individual comb trees and will
(for the most part) only change if their comb does.  The exception is
valid comb instances becoming invalid or vice versa if objects are
added/ deleted, but there too the nref tree isn't really necessary.