Stash work on testing tinygl as a libdm backend.
Didn't take this all the way to working, but got far enough to see that
tinygl doesn't have all the pieces our current libdm OpenGL code is
using.  We could undoubtedly get it there if need be, by grabbing the
pieces we need from osmesa, but seeing as osmesa is working and gives us
a more complete OpenGL stack it's probably the way to go.

Stashing work done so far on tinygl as a patch - if we decide at some
point we want to give tinygl another look, this commit will preserve the
initial steps taken.