Replace -idn search opt with more general -matrix
Per suggestion from Sean, replace the -idn option (which was limited to
identity matrices) with a more general -matrix filter.  IDN is supported
as a way to specify the identity matrix ("-idn" is now "-matrix IDN")
but the new logic also allows specification of a full 16 number matrix
as an alternative.