Shift Eigen from src/other to src/other/ext
src/other/ext is intended to hold copies of third party dependencies that
meet the following criteria:

a) Upstream is active
b) It is plausible that we can use an unaltered system install of the
   upstream dependency for our needs.

Only those deps which fail to satisfy one or both of those criteria are
stored directly in src/other.  Eigen satisfies the criteria, so shift it
down into ext.