address slew of new type mismatches gcc 11 detects
lastest gcc is much better at tracking typedef types and noticing
when functions are declared with one type and defined with other
type-compatible parameters.

used to be a problem long ago with some compilers not working well
with typedef types as parameters (e.g., vect_t vs fastf_t *) but
this forunately/hopefully is a vestige.  updates tested on linux
and mac, so hopefully windows behaves.