add a donuts procedural database generator that makes torii patterns
this is from over 15 years ago, and wasn't the final version used to
model the stacked torii demo, but was the only version seemingly
preserved.  it needs a LOT more tlc, but is finally at least compiling
(mac).  still needs de-k&r'ing, header, footer, and an option to not
fan out.