Going to need some way to test QgInstance et. al.
There are going to be a lot of pieces to the Qt .g model, and we need
some way to incrementally test small parts of the logic without the
complexity of standing up the all-up Qt widget.  Add in a simple program
to read in a .g file and exercise the data structures.  Eventually
pieces of this may mature enough to migrate either to libraries or qged,
but for the moment its scope is just "development tool".