Don't eliminate children from items on close
Archer's tree remembers what subtrees were open and closed if a higher
level object is opened or closed - this is desirable behavior, so think
about how it might be achieved within the Qt context.  Basically we
don't want to wipe out pre-existing QgItems on close - keep them, just
don't display them or their subtrees.  It's only when one is found to be
invalid that it and its subtrees need to go.