Set up to supply a dbip pointer in struct ged
Of 1490 instances matching the pattern gedp->ged_wdbp, 1150 of them
appear to be doing so to get the dbip pointer from wdbp.  Also, if a ged
command isn't going to work with the rt_wdb structure, we shouldn't
really be requiring it for other commands - i.e. a ged_init without a
wdbp should still enable as much of the libged logic as can be used with
just a dbip, and fail cleanly for functionality requiring a non-NULL
wdbp.  Set the stage for some experiments.