clear out stale nirt issues.
the issue with changing units is a deep and not likely to get fixed anytime soon.
was unable to reproduce the nirt -e parsing issue inside/outside mged. perhaps
it's platform specific.  the sf bug report need not be replicated here, and that
issue hasn't been observed.  uninstalled behavior is a nicety, but not necessary.