Update fnmatch code to the latest from OpenBSD
While we're at it, also hook up the regressions tests from their repo to
test the functionality on an ongoing basis.  Reworked bu_test a bit in
the process to be a C++ app, so we can use either C or C++ in setting up
tests.  UUID code really doesn't get along with C++, so give it its own
executable.