Don't static analyze obj re2c generated code
There isn't really a good solution for this that I can see - even if
there is a theoretical way to update re2c to avoid dead assignment
issues (and at this point I'm not convinced there is) it's not practical
from a resource availability standpoint to try at this point.  To avoid
false positives, since we eventually want to get the build to a point
where it is clean and we can enforce keeping it that way, use a clang
mechanism to avoid the errors from the generated file.  I don't like it,
and if clang adds a way in the future to do more targeted disablement of
just dead store reporting we should use that, but in the meantime this
will avoid producing reports from the machine generated C++ file.