Check found before returning -1.
This one has potential implications and needs testing.  Logic added in
r60186 adds a fallback test to look at pt instead of from, but as
implemented despite doing checks for within-tol-but-not-at
BREP_SAME_POINT_TOLERANCE in the latter stage it only reports success if
the result actually IS within BREP_SAME_POINT_TOLERANCE.  This change
mirrors the behavior of the "from" test above, and may report success
where previously we would not have.  Don't know yet what (if anything)
that might change.