rename v4 material record id from ID_MATERIAL to ID_COLORTAB
this is to make room for new v5 'material' database objects being implemented
by a team of students at TAMU.  this change is minimally impacting as a simple
rename, but is also only affecting v4 API that has now been out of date for a
very long time.  moreover, the new label is more reflective of the data held
by the record as it's color table information defined by ident ranges and rgb
values.