Make a stab at consolidating memory lookup code
GCT has some code that looks up available memory and page size.  We
needed similar information for facetize, and the beginnings of a libbu
capability to provide that information were started a few years back.
See if we can consolidate the two.

This is unfortunately one of those cases where the only APIs that
provide the info we need are platform specific, and the terminology
isn't uniform between systems either.  This will probably need tweaking
and refinement over time, but at least get the ball rolling.