Reorganize a bit to make the analyzer logic clear
This one is a puzzler...  Even when I reorganize the logic as in this
commit, the analyzer is assuming end_orad' is <= field 'dist' at line
2567 and then immediately turning around as assuming 'end_orad' is >
field 'dist' on line 2572.  end_orad is passed in as a number and tol is
const - the only thing I can think of is maybe the analyzer is allowing
for the possibility that something else (another thread?) is altering
tol->dist independent of this function?