Reduce the r50164 additions to their core
OK, it's getting clearer.  The key point of the additional code is to
satisfy subsequent filtering code by setting hit_vpriv[X], and to do
so a block of code was copied over to do so.  However, with the
exception of hit_vpriv[X] itself, which has other inputs being used to
set it, the other calculations all depend on unset values.

See if we can strip this down to its current core functionality.  Not
sure this is where it should ultimately land (ideally we probably want
all these values to be meaningfully set, if TIE allows for that, since
other functions do seem to reference them) but at least we want to avoid
the impression of meaningful number assignments when we don't have the
input values to back them.