Start looking at libbrep null ref returns
Clang static analyzer report for libbrep has a number of rather obtuse
"Returning null reference" reports.  These usually report a reference
coming from from opennurbs_array_defs.h, but after some digging the
problems generally seem to be with our initialization and (lack of)
error handling in libbrep itself.

This is probably going to be one of those situations where addressing
one issue (by, say, bailing with an error condition instead of
continuing with invalid inputs) will propagate to other issues (i.e. a
caller now needs to handle the error condition it was previously
ignoring.)  No help for it, but the result is that it may be a while
before this work really begins to consistently drive down the static
analyzer error count.