Remove the naca experimental port to C
Generating some clang static analyzer errors - just removing as the
simplest step.  Probably not too hard to fix, but we never got this to
the point where we were able to generate geometry and the Vehicle Sketch
Pad software (OpenVSP, http://openvsp.org/) originally developed by NASA
has the ability to generate STEP files.  Unless someone wants to pick
this up (in which case the VCS has it) the better investment of effort
is to  firm up our step-g importer's ability to bring in OpenVSP
exports.