Simplify static analyzer script
With more experimentation and shrinking error count, the simpler thing
to do looks to be to just use the "all" target after the libraries are
built to collectively check the remaining logic.

If we can get to a fully clean build, we'll probably be able to simplify
much further and dispense with the preliminary library building stage -
just doing an "all" build after the src/other and misc/tools components
are built should return a clean build, and if not we are unlikely to
need to break the outputs out into individual per-library reports.