ON_SimpleArray -> ON_ClassArray
Sean spotted this, and it looks like it may have been a significant
systematic mistake in my early NURBS brep coding.  ON_SimpleArray can't
always be used - per opennurbs_array.h, ON_SimpleArray "should not be
used for arrays of classes that require explicit construction,
destruction, or copy operators."  That probably means we've got a lot of
inappropriate uses of ON_SimpleArray at play.  I tended to try and treat
pointers to ON_* objects as being inserted into and managed by
ON_SimpleArray essentially as if they were C arrays of pointers, and
that looks to be incorrect.  Good chance this is the root cause behind a
lot of the more obscure memory issues clang static analyzer is finding
with the various brep related codes...