Watch analyzer results for osmesa, utahrle, polyt2ri
Having gotten these clean (at least with clang 12), lets keep them that
way.  utahrle doesn't have an active upstream, our poly2tri is
significantly altered from the upstream version, and osmesa is an
extraction of Mesa 7.0.4's OpenGL 2.1 software-only rasterizer - the
current upstream code is now completely different.

Realistically Tktable and Tkhtml (particularly Tkhtml) should also fall
into this category, but they're not as trivial to set up in the test
script since they may need Tcl/Tk built from src/other/ext first (and,
unlike the other three libs, the Tk widgets should eventually be phased
out in favor of Qt functionality anyway...)

OpenNURBS is in a separate class - basically, unlike the other src/other
libs not in ext, it does have an active upstream - we just need to
maintain it in src/other because of the significant number of patches we
apply to it.