Try adding sentinel files to the perplex/lemon cpy
Grr.  Getting random failures in the stepcode building, mostly on the
Github runners but occasionally on normal desktops as well.  Symptoms
are an internal operation of the stepcode build erring out complaining
about not being able to do I/O on one of the files that needs to be
copied, but when you check the file is there.  Historically that's
happened when disk I/O isn't quite keeping up with processing or
parallel operations end up colliding.

Traditional solution has been to add sentinel files that are only
created after the original operation is done.  See if that helps here.