Simplify static analysis script
With the count down, it's no longer as beneficial to separate things out
into per-library reports - just assemble a single report for the all
target, and use the total report count as a criteria for failure - like
our other checks, if the count rises over the current known rate report
failure.  We'll lower the number as issues are resolved.