Sanity check yyruleno value
CID 1127262.  If I'm understanding this correctly, Coverity is spotting
the bounds check in the debug printing statement, interpreting this as
an indication that yyruleno might exceed the array size, and (assuming
that's true) then points out that if yyruleno CAN do that it would
result in a problem when the array contents are accessed later on.

Just do the check up front, to make sure we don't get into this logic
with a problematic yyruleno, and warn if we hit it.  Not sure if the
debugging logic was doing this check on general principles or they had
an issue at some point...