Use the %p specifier (standard) to print pointers
CID 1488175 is indicating pc->pkc_inbuf isn't NULL terminated because
it's contents come from read (which is true) and therefore snprintf has
a problem with it.  That would be true if we were trying to print the
pkc_inbuf contents as a string, but I think the intent here is to print
pointers, which should make NULL termination moot.  (Printing the
pointers here is of questionable utility, but that's another
discussion...)  See if using %p is more acceptable to Coverity.