Test Coverity's reaction to a refactor
Having trouble understanding why g2asc.c is triggering CID 1488403.  I'm
wondering if the Coverity scanner isn't properly unwinding
BU_LIST_WHILE.  That macro is doing a new assignment of headp->forw to
mp, but if Coverity is instead interpreting BU_LIST_WHILE as a function
call that uses the just-freed mp value it might explain why it looks
like a use-after-free error to the scanner.  This probably isn't a
general solution, but in this case its easy enough to refactor to not
use the macro - let's see if another scan still reports the issue.