Start working on cyclic path reporting in librt
The libged lint functionality has the ability to identify and report
cyclic paths.  Going to need some form of that for the libqtcad .g model
(if it can be done with a fast enough performance) so start thinking
about shifting it down into librt.  Haven't refactored the ged logic to
use this yet - need to set up some unit tests and check new logic
against the old version before removing it.