Adjust position tracking in strnstr
Coverity spotted some logically dead code - hpos starts as zero, and the
way the for loop was set up that resulted in a failed test that
immediately halted the for loop.  Adjust to do the increment inside the
loop body, rather than in a location where the hpos value itself acts as
a condition test.

CID 1488221, 1488445, 1488991