Experiment with suppressing a Coverity warning
CID 1488777 reports "bad_sizeof: Taking the size of &internal, which is
the address of an object, is suspicious" for RT_CK_DB_INTERNAL(&internal)
in the libgcv plugin file decimate/decimate.c:60  I think(?) this is
the intended logic here, so see if we can tell Coverity that.

Better practice would be to make some sort of model entry to cover this
case... not sure if the model mechanism covers this case yet, but if we
can figure that out this comment should be removed in favor of a model
based solution.