Add a place to hang basic bu_list tests.
There are a lot of Coverity reports that seem to pertain to bu_list
behavior, but in most cases it's difficult to follow the logic - the
callback stacks can be rather deep.  Set up a simple bu_test harness to
poke and prod various loop behaviors to see if we can identify what the
problems might be in a simpler context.