Alter clear_list memory cleanup for GED check cmd
This one is a legit crasher.  The BU_LIST_FOR iteration can't do the
clean up of the bu_list containers itself (at least, not this way.) Just
do the cleanup of the non-bu_list-container memory with the for loop
(which bu_list_free can't do), and then let bu_list_free take care of
the more finicky cleanup needed for the bu_list containers themselves.

CID 1489153