Checkpoint some work on an NMG use of bu_list
Coverity doesn't like the memory behavior around libnmg/mk.c1558.
Trying to understand what is happening and what needs to happen when -
this may require some unit testing of libnmg functions so we can poke
the right logic in a self-contained fashion.  These changes don't break
the "make check" target on Linux, but that's not saying a whole lot...
Not at all convinced this is correct or properly freeing memory yet
(probably not, in fact) - more to come.

CID 1489028