Note issues with removing NMG attr structs
Make a quick try at folding the shell bbox min and max points into the
shell container and removing shell_a.  For NMG itself it looks doable,
but librt defines export-to-disk logic that specifically encodes shell_a
containers - this implies that the on-disk format itself requires these
structures.

If possible, need to get the serialization/deserialization code out of
librt and into libnmg (have it encode/decode to bu_externals or whatever
container is appropriate, then hand back to librt for the .g database
part...)