Experiment with moving NMG I/O to libnmg
The serialization and de-serialization of NMG data structures is quite
tightly coupled to the in-memory storage.  To properly decouple librt
and libnmg, we need to move the parts of this logic that are aware of
the internal NMG details out of librt - librt's role would then be
limited to the bu_external; pass it in to libnmg for import, and write
out the bu_external provided by libnmg for export.